Preparing for distribution.
I registered for PyPI account and have modified the source to be
compliant with packaging for PyPI. In fact, the application is now
uploaded! Still need to test it though. Brendan checked and mentioned
that it's not 'runnable'. Need to figure out if that is a user-side
setting or if I need to do something different in order for the package
to work. But I think I do have partial success here.

First moved listmaker sources into a subdirectory inside listmaker. This
seems to be the way most developer do things so following suit. Added
files required for packaging.

Command I used was:
$ python3 setup.py sdist upload